restructure, button cooldown
