+++ webapps/shared/nginx/nginx.conf
@@ -3,6 +3,31 @@ events {}
 http {
     resolver 127.0.0.11 valid=10s; 
 
+        server {
+            listen 80;
+            server_name template.patrickdd.com;
+
+        location /api/ {
+            set $upstream_template_backend http://template-backend:8080;
+
+            proxy_pass $upstream_template_backend;
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            proxy_set_header X-Forwarded-Proto $scheme;
+        }
+
+        location / {
+            set $upstream_template_frontend http://template-frontend:80;
+
+            proxy_pass $upstream_template_backend;
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            proxy_set_header X-Forwarded-Proto $scheme;
+        }
+    }
+
     server {
         listen 80;
         server_name patrickdd.com, patrickdumdum.com;
@@ -23,6 +48,7 @@ http {
 
         location / {
             set $upstream_fsp_frontend http://fsp-frontend:80;
+
             proxy_pass $upstream_fsp_frontend;
             proxy_set_header Host $host;
             proxy_set_header X-Real-IP $remote_addr;
@@ -32,6 +58,7 @@ http {
 
         location /api/ {
             set $upstream_fsp_backend http://fsp-backend:8080;
+            
             proxy_pass $upstream_fsp_backend;
             proxy_set_header Host $host;
             proxy_set_header X-Real-IP $remote_addr;
