+++ webapps/template/frontend/src/app/app.routes.ts
@@ -1,6 +1,6 @@
 import { Routes } from '@angular/router';
-import { HomeComponent } from './components/home/home.component';
-import { PingComponent } from './components/ping/ping.component';
+import { HomeComponent } from './features/home/components/home.component';
+import { PingComponent } from './features/ping/components/ping.component';
 
 export const routes: Routes = [
     { path: '', component: HomeComponent },
