--- webapps/template/frontend/src/app/components/ping/ping.component.html
@@ -1,1 +1,0 @@
-<button (click)="sendPing()">Ping Backend</button>
