+++ webapps/template/frontend/src/app/core/services/api.service.spec.ts
@@ -1,5 +1,4 @@
 import { TestBed } from '@angular/core/testing';
-
 import { ApiService } from './api.service';
 
 describe('ApiService', () => {
