+++ webapps/template/frontend/src/app/features/ping/components/ping.component.html
@@ -1,0 +1,2 @@
+<button [disabled]="buttonDisabled()" (click)="sendPing()">Ping Backend</button>
+<p *ngIf="pingResponse">{{ pingResponse() }}</p>
