+++ webapps/template/frontend/src/app/features/ping/components/ping.component.spec.ts
